Moved helper functions to utils
